backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
aiohttp
elevenlabs
pydantic-settings
